requirements.txt
langchain
ollama
faiss-cpu
sentence-transformers
spacy
